+++ soya/HISTORY.md
@@ -29,6 +29,7 @@
   - ReduxStore.register() also register Segment dependencies recursively.
   - ReduxStore.dispatch() process QueryDependencies.
   - Recursive QueryDependencies tested to be working.
+  - Function based QueryDependencies tested to be working.
 
 ## 0.0.23